+++ src/AtomUI.Desktop.Controls/Window/Chrome/WaylandWindowChromeManager.cs
@@ -1,4 +1,5 @@
 using System.Runtime.Versioning;
+using AtomUI.Native;
 using Avalonia;
 using Avalonia.Controls;