+++ src/AtomUI.Desktop.Controls/Window/Utils/FullscreenPopoverLayer.cs
@@ -1,5 +1,4 @@
 using System.Reactive.Disposables;
-using AtomUI.Theme;
 using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Controls.Primitives;